Add assembly definition + organize struct folder structure
